-Removed all draggability and set the resize event to reinitialize the simulation -Removed all initial attemps at ghosting (to be replaced with a ghosting engine which supports multiple ghosts) -Reorganized the file structure to be standard with new style guidelines -Made div and canvas dynamically created in javascript for compactness -Reorganized some basic variables and manipulated a few functions for modularity
NOTE: This is a great working version with no known bugs that shows off the basic time travel elements.
Other than some simple style changes, modularity cleaning and reorganization, this version is very stable.
The only psudo-bug noted is that if the simulation becomes horribly lagged or paused it will play "catch up". This does not throw the simulation off and is simply a visual difference. Might still be worth looking into fixing.

Signed-off-by: unknown <admin@admin-NB314.(none)>